feat: user dataProvider & user /me route & children wrapped with Auth provider

+++ src/app/api/movies/[id]/route.js
@@ -1,6 +1,7 @@
 import connectDB from "src/lib/mongodb";
 import Movie from "src/models/model.movies.js";
 import { NextResponse } from "next/server";
+import { checkAuth } from "src/lib/auth";
 
 export const GET = async (req, { params }) => {
   const id = await params.id;
@@ -14,28 +15,58 @@ export const GET = async (req, { params }) => {
 };
 
 export const DELETE = async (req, { params }) => {
+  await connectDB();
+  const authenticatedUser = await checkAuth(req);
+
+  if (!authenticatedUser) {
+    return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+  }
+
+  const isAdmin = authenticatedUser.role === "admin";
+
+  if (!isAdmin) {
+    return NextResponse.json(
+      { message: "You dont have the right to use this feature!" },
+      { status: 403 }
+    );
+  }
+
   try {
     const id = params.id;
-    await connectDB();
 
     await Movie.findByIdAndDelete(id);
-  }catch(error){
+  } catch (error) {
     console.error(error);
   }
 };
 
 export const PUT = async (req, { params }) => {
-    const id = params.id;
-    const body = await req.json();
-    const { inCinemas } = body;
+  await connectDB();
+  const authenticatedUser = await checkAuth(req);
+  console.log(authenticatedUser);
 
-    await connectDB();
+  if (!authenticatedUser) {
+    return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+  }
 
-    const updatedMovie = await Movie.findByIdAndUpdate(
-      id,
-      { inCinemas },
-      { new: true, runValidators: true }
+  const isAdmin = authenticatedUser.role === "admin";
+
+  if (!isAdmin) {
+    return NextResponse.json(
+      { message: "You dont have the right to use this feature!" },
+      { status: 403 }
     );
+  }
 
-    return NextResponse.json(updatedMovie, { status: 200 });
+  const id = params.id;
+  const body = await req.json();
+  const { inCinemas } = body;
+
+  const updatedMovie = await Movie.findByIdAndUpdate(
+    id,
+    { inCinemas },
+    { new: true, runValidators: true }
+  );
+
+  return NextResponse.json(updatedMovie, { status: 200 });
 };
